created some scope to gather assignments to be considered for reminders
